refactor(jans-pycloudlib): share storage API for config, secret, and lock
